# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Usuários",
    "version": "1.0.0",
    "description": "API para registro, login e consulta de usuários."
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido. Retorna o token JWT." },
          "400": { "description": "Credenciais inválidas ou dados ausentes." }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Listar usuários",
        "responses": {
          "200": { "description": "Lista de usuários." }
        }
      },
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "favored": { "type": "boolean" },
                  "balance": { "type": "number" }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado com sucesso." },
          "400": { "description": "Erro de validação ou negócio." }
        }
      }
    },
    "/transfers": {
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Realizar transferência entre usuários",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fromId": { "type": "string" },
                  "toId": { "type": "string" },
                  "amount": { "type": "number" }
                },
                "required": ["fromId", "toId", "amount"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transferência realizada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "from": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "balance": { "type": "number" }
                      }
                    },
                    "to": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "balance": { "type": "number" }
                      }
                    },
                    "transfer": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "fromId": { "type": "string" },
                        "toId": { "type": "string" },
                        "amount": { "type": "number" },
                        "createdAt": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação ou negócio.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Usuário remetente ou destinatário não foram encontrados" }
                  }
                }
              }
            }
          },
          "401": { "description": "Token não fornecido ou inválido." } 
        }
      },
      "get": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Listar transferências",
        "responses": {
          "200": {
            "description": "Lista de transferências.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "fromId": { "type": "string" },
                      "toId": { "type": "string" },
                      "amount": { "type": "number" },
                      "createdAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Token não fornecido ou inválido." }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}